Add WarnChannels as a component
