+++ api/web/src/components/CloudTAK/Map.vue
@@ -27,6 +27,11 @@
             />
         </div>
 
+        <WarnChannels
+            v-if='warnChannels'
+            @close='warnChannels = false'
+        />
+
         <div
             v-if='profile'
             class='position-absolute bottom-0 begin-0 text-white'
@@ -372,54 +377,15 @@
                 </div>
             </TablerModal>
         </template>
-
-        <template v-if='warnChannels'>
-            <TablerModal>
-                <div class='modal-status bg-red' />
-                <button
-                    type='button'
-                    class='btn-close'
-                    aria-label='Close'
-                    @click='warnChannels = false'
-                />
-                <div class='modal-header text-white'>
-                    <div class='d-flex align-items-center'>
-                        <IconInfoSquare
-                            :size='28'
-                            :stroke='1'
-                        />
-                        <span class='mx-2'>No Channels Selected</span>
-                    </div>
-                </div>
-                <div class='modal-body text-white'>
-                    <p>Channels are the basis of who and what you are able to see in TAK</p>
-
-                    <p>At this time you do not have any active channels and as such you will not see any Users, Data Packages, or Data Syncs associated with channels.</p>
-
-                    <p>It is recommended to turn on relevant channels now</p>
-
-                    <div class='d-flex align-items-center'>
-                        <div class='ms-auto btn-list'>
-                            <button
-                                class='btn btn-primary'
-                                @click='selectChannels'
-                            >
-                                Select Channels
-                            </button>
-                        </div>
-                    </div>
-                </div>
-            </TablerModal>
-        </template>
     </div>
 </template>
 
 <script>
+import WarnChannels from './util/WarnChannels.vue';
 import { std, stdurl } from '/src/std.ts';
 import CloudTAKFeatView from './FeatView.vue';
 import {
     IconSearch,
-    IconInfoSquare,
     IconMessage,
     IconLocationOff,
     IconLocation,
@@ -602,10 +568,6 @@ export default {
             this.feat = false;
             this.$router.push("/");
             this.pointInput.shown = false;
-        },
-        selectChannels: function() {
-            this.warnChannels = false;
-            this.$router.push('/menu/channels');
         },
         submitPoint: async function() {
             this.pointInput.shown = false;
@@ -902,6 +864,7 @@ export default {
         }
     },
     components: {
+        WarnChannels,
         SideMenu,
         Loading,
         SelectFeats,
@@ -917,7 +880,6 @@ export default {
         IconMessage,
         IconLocationOff,
         IconLocation,
-        IconInfoSquare,
         IconMinus,
         IconBell,
         IconPlus,
